pyproject.toml
[project]
name = "biochem-agent"
version = "0.1.0"
description = "Agentic biochemical data analysis assistant using Chainlit, OpenAI, and Tavily."
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "chainlit>=2.4.400",
    "openai>=1.77.0",
    "tavily-python>=0.7.12",
    "pandas>=2.2.3",
    "numpy>=2.2.4",
    "matplotlib>=3.9.0",
    "seaborn>=0.13.2",
    "plotly>=6.0.1",
    "pydantic>=2.6.0",
    "Pillow>=10.0.0",
    "httpx>=0.28.1",
    "requests>=2.31.0",
    "python-dotenv>=1.1.0",
    "scipy>=1.15.2"
]
